+++ packages/frameworks/vue/CHANGELOG.md
@@ -9,6 +9,7 @@ description: All notable changes to this project will be documented in this file
 ### Added
 
 - Added `DatePicker` component
+- Added `Progress` component
 - Added `modelValue` for `Dialog` component
 - Added `modelValue` for `HoverCard` component
 - Added `modelValue` for `Popover` component